// modules/home/models/AppSettings.ts
export interface AppSettings {
  whatsappNumber: string;
  contactEmail: string;
  whatsappGeneralMessage: string;
  whatsappProductMessage: string;
  whatsappFABMessage: string;
  shippingTimeLima: string;
  shippingTimeProvincia: string;
  galleryImages: string[];
  heroTitle: string;
  heroTitleHighlight: string;
  heroSubtitle: string;
  heroImageUrl: string;
  heroWhatsappText: string;
  heroBadgeText: string;
  heroSubBadge: string;
}